+++ lib/_tls_common.js
@@ -56,7 +56,7 @@ const {
 
 const {
   parseCertString,
-} = require('internal/tls/legacy');
+} = require('internal/tls/parse-cert-string');
 
 function toV(which, v, def) {
   if (v == null) v = def;
