+++ src/node_native_module.cc
@@ -99,7 +99,8 @@ void NativeModuleLoader::InitializeModuleCategories() {
       "tls",
       "_tls_common",
       "_tls_wrap",
-      "internal/tls/legacy",
+      "internal/tls/secure-pair",
+      "internal/tls/parse-cert-string",
       "internal/tls/secure-context",
       "internal/http2/core",
       "internal/http2/compat",
